Moving files to the proper place + create map behaviour
