+++ lib/assets/javascripts/cartodb3/dataset.js
@@ -15,8 +15,8 @@ var _ = require('underscore');
 var Backbone = require('backbone');
 var ConfigModel = require('./data/config-model');
 var TableModel = require('./data/table-model');
-var DatasetView = require('./dataset/dataset-content/dataset-view');
-var OptionsView = require('./dataset/dataset-options/dataset-options-view');
+var DatasetContentView = require('./dataset/dataset-content/dataset-content-view');
+var DatasetOptionsView = require('./dataset/dataset-options/dataset-options-view');
 var QuerySchemaModel = require('./data/query-schema-model');
 var ModalsServiceModel = require('./components/modals/modals-service-model');
 var SynchronizationModel = require('./data/synchronization-model');
@@ -78,7 +78,7 @@ var headerView = new HeaderView({
 });
 $('.js-info').append(headerView.render().el);
 
-var datasetView = new DatasetView({
+var datasetContentView = new DatasetContentView({
   el: $('.js-table'),
   modals: modals,
   userModel: userModel,
@@ -87,9 +87,9 @@ var datasetView = new DatasetView({
   configModel: configModel,
   querySchemaModel: querySchemaModel
 });
-datasetView.render();
+datasetContentView.render();
 
-var optionsView = new OptionsView({
+var datasetOptionsView = new DatasetOptionsView({
   el: $('.js-datasetOptions'),
   router: Router,
   modals: modals,
@@ -100,7 +100,7 @@ var optionsView = new OptionsView({
   syncModel: syncModel,
   querySchemaModel: querySchemaModel
 });
-optionsView.render();
+datasetOptionsView.render();
 
 var notifierView = Notifier.getView();
 $('.js-notifier').append(notifierView.render().el);
