+++ lib/assets/javascripts/cartodb3/dataset/dataset-content/dataset-content-options-view.js
@@ -1,5 +1,5 @@
 var CoreView = require('backbone/core-view');
-var template = require('./dataset-options.tpl');
+var template = require('./dataset-content-options.tpl');
 
 module.exports = CoreView.extend({
