+++ lib/assets/javascripts/cartodb3/dataset/dataset-content/dataset-content-options.tpl
