+++ lib/assets/javascripts/cartodb3/dataset/dataset-content/dataset-content-view.js
@@ -1,6 +1,6 @@
 var CoreView = require('backbone/core-view');
 var TableManager = require('../../components/table/table-manager');
-var DatasetOptionsView = require('./dataset-options-view');
+var DatasetContentOptionsView = require('./dataset-content-options-view');
 
 module.exports = CoreView.extend({
 
@@ -55,7 +55,7 @@ module.exports = CoreView.extend({
     this.addView(tableView);
     this.$el.append(tableView.render().el);
 
-    var datasetOptionsView = new DatasetOptionsView({
+    var datasetContentOptionsView = new DatasetContentOptionsView({
       modals: this._modals,
       userModel: this._userModel,
       tableModel: this._tableModel,
@@ -63,15 +63,15 @@ module.exports = CoreView.extend({
       syncModel: this._syncModel,
       configModel: this._configModel
     });
-    datasetOptionsView.bind('addColumn', function () {
+    datasetContentOptionsView.bind('addColumn', function () {
       tableView.addColumn();
     });
-    datasetOptionsView.bind('addRow', function () {
+    datasetContentOptionsView.bind('addRow', function () {
       tableView.addRow();
     });
 
-    this.addView(datasetOptionsView);
-    this.$el.prepend(datasetOptionsView.render().el);
+    this.addView(datasetContentOptionsView);
+    this.$el.prepend(datasetContentOptionsView.render().el);
   },
 
   _onExportClick: function () {
