+++ lib/assets/javascripts/cartodb3/dataset/dataset-options/dataset-options-view.js
@@ -1,8 +1,9 @@
 var CoreView = require('backbone/core-view');
 var template = require('./dataset-options.tpl');
+var VisDefinitionModel = require('../../data/vis-definition-model');
 var CreationModalView = require('../../components/modals/creation/modal-creation-view');
-var TITLE_SUFFIX = ' | CartoDB';
-var DATASET_URL_PARAMETER = '/dataset/';
+var errorParser = require('../../helpers/error-parser');
+var MAP_URL_PARAMETER = '/builder/';
 
 module.exports = CoreView.extend({
 
@@ -49,21 +50,22 @@ module.exports = CoreView.extend({
         loadingTitle: _t('dataset.create-map.loading', { tableName: tableName }),
         errorTitle: _t('dataset.create-map.error', { tableName: tableName }),
         runAction: function (opts) {
-          console.log('running action');
-          // tableDuplicationOperation({
-          //   query: self._querySchemaModel.get('query'),
-          //   tableModel: self._tableModel,
-          //   configModel: self._configModel,
-          //   onSuccess: function (importModel) {
-          //     var tableName = importModel.get('table_name');
-          //     window.location = self._configModel.get('base_url') + DATASET_URL_PARAMETER + tableName;
-          //   },
-          //   onError: function (importModel) {
-          //     var error = importModel.get('get_error_text');
-          //     var errorMessage = error && error.title;
-          //     opts.error && opts.error(errorMessage);
-          //   }
-          // });
+          var newVisModel = new VisDefinitionModel({
+            name: self._visModel.get('name') + ' ' + _t('editor.map')
+          }, {
+            configModel: self._configModel
+          });
+
+          newVisModel.save({
+            source_visualization_id: self._visModel.get('id')
+          }, {
+            success: function (visModel) {
+              window.location = self._configModel.get('base_url') + MAP_URL_PARAMETER + visModel.get('id');
+            },
+            error: function (mdl, e) {
+              opts.error && opts.error(errorParser(e));
+            }
+          });
         }
       });
     });
