+++ lib/assets/javascripts/cartodb3/dataset/dataset-options/dataset-options.tpl
@@ -2,7 +2,7 @@
   <li class="CDB-InfoBox-footerItem js-controls"></li>
   <li class="CDB-InfoBox-footerItem CDB-InfoBox-footerItem--right js-actions">
     <button class="CDB-Button CDB-Button--primary CDB-Button--small u-upperCase js-createMap">
-      <span class="CDB-Text is-semibold CDB-Size-small"><%- _t('dataset.create-map') %></span>
+      <span class="CDB-Text is-semibold CDB-Size-small"><%- _t('dataset.create-map.title') %></span>
     </button>
   </li>
 </ul>
