+++ lib/assets/test/spec/cartodb3/dataset/dataset-view.spec.js
@@ -1,7 +1,7 @@
 var Backbone = require('backbone');
 var _ = require('underscore');
 var ConfigModel = require('../../../../javascripts/cartodb3/data/config-model');
-var DatasetView = require('../../../../javascripts/cartodb3/dataset/dataset-content/dataset-view');
+var DatasetView = require('../../../../javascripts/cartodb3/dataset/dataset-content/dataset-content-view');
 var UserModel = require('../../../../javascripts/cartodb3/data/user-model');
 var TableModel = require('../../../../javascripts/cartodb3/data/table-model');
 var QuerySchemaModel = require('../../../../javascripts/cartodb3/data/query-schema-model');
